Fix #663, unit test support functions
Update use of uint32 to size_t in UT support code

+++ fsw/cfe-core/unit-test/ut_support.h
@@ -572,7 +572,7 @@ uint16 UT_GetNumEventsSent(void);
 **        This function does not return a value.
 **
 ******************************************************************************/
-void UT_DisplayPkt(CFE_SB_MsgPtr_t ptr, uint32 size);
+void UT_DisplayPkt(CFE_SB_MsgPtr_t ptr, size_t size);
 
 /*****************************************************************************/
 /**
